Browse part name dialog opens at saved directory, and loads part name.  Part name and dir are saved also
Possible issue: Sometimes the overwrite dialog is brought up unnecessarily. If the user were to type in the name of a non .dat file (ex: 'brick.stl') which already exists, the overwrite prompt would come up. However, the file is not really at risk of being overwritten, because '.dat' is appended at the end, making the filename 'brick.stl.dat'